fix: build the cats matchers for the JVM only
IOMatchers calls unsafeRunSync, which does not exist in the Scala.js and
Scala Native builds of cats-effect, so those modules were compiled against the
JVM artifact instead. The platform independent matchers move to
RunTimedMatchers and IOMatchers becomes a JVM source, which lets cats resolve
its own artifact on each platform.
