+++ build.sbt
@@ -101,10 +101,9 @@ lazy val analysisNative = analysis.native
 lazy val cats = crossProject(JSPlatform, JVMPlatform, NativePlatform).in(file("cats")).
   settings(
     commonSettings,
-    // the JVM artifacts are used on every platform, as they were with sbt 1
     libraryDependencies ++= Seq(
-      ("org.typelevel" %% "cats-core" % catsVersion).withPlatformOpt(Some("jvm")),
-      ("org.typelevel" %% "cats-effect" % catsEffectVersion).withPlatformOpt(Some("jvm"))
+      "org.typelevel" %% "cats-core" % catsVersion,
+      "org.typelevel" %% "cats-effect" % catsEffectVersion
     ),
     name := "specs2-cats"
   ).
