+++ cats/shared/src/main/scala/org/specs2/matcher/RunTimedMatchers.scala
@@ -1,6 +1,5 @@
 package org.specs2.matcher
 
-import cats.effect.IO
 import org.specs2.matcher.ValueChecks.valueIsTypedValueCheck
 import org.specs2.matcher.describe.Diffable
 
@@ -70,23 +69,4 @@ trait RunTimedMatchers[F[_]] {
     def withValue(t: T)(implicit di: Diffable[T]): TimedMatcher[T] =
       timedMatcher.withValue(valueIsTypedValueCheck(t))
   }
-}
-
-trait IOMatchers extends RunTimedMatchers[IO] {
-
-  import cats.effect.unsafe.implicits.global
-
-  protected def runWithTimeout[A](fa: IO[A], d: FiniteDuration): A = fa.timeout(d).unsafeRunSync
-  protected def runAwait[A](fa: IO[A]) : A = fa.unsafeRunSync
-
-  protected[specs2] override def attemptRun[T](check: ValueCheck[T], duration: Option[FiniteDuration]): IOMatcher[T] =
-    IOMatcher(check, duration)
-
-  case class IOMatcher[T](check: ValueCheck[T], duration: Option[FiniteDuration])
-      extends TimedMatcher[T](check, duration) {
-    def checkIOWithDuration[S <: IO[T]](e: Expectable[S], d: FiniteDuration): MatchResult[S] =
-      checkWithDuration(e, d)
-    def checkIO[S <: IO[T]](e: Expectable[S]) = checkAwait(e)
-  }
-
 }
